2019 - Day 08a

+++ src/day_template.rs
@@ -1,8 +1,8 @@
-pub fn run_a(input: &Vec<String>) -> i64{
+pub fn run_a(input: &Vec<String>) -> i32 {
     0
 }
 
-pub fn run_b(input: &Vec<String>) -> i64 {
+pub fn run_b(input: &Vec<String>) -> i32 {
     0
 }
